+++ catkin_ws/src/telemetry/src/transceiver/transceiver.py
@@ -1,0 +1,76 @@
+# File: telemetry/src/transceiver/transceiver.py
+import json
+import sys
+import termios
+from catkin_ws.src.telemetry.src.transceiver._sx126x import sx126x
+from telemetry.msg import command_msg
+
+class Transceiver(sx126x):
+    def __init__(
+            self, 
+            serial_num="/dev/ttyS0", 
+            freq=433, 
+            addr=0, 
+            power=22, 
+            rssi=False, 
+            air_speed=2400, 
+            relay=False
+        ) -> None:
+        super().__init__(serial_num=serial_num, freq=freq, addr=addr, power=power, rssi=rssi, air_speed=air_speed, relay=relay)
+        
+        # Terminal settings
+        self.old_settings = termios.tcgetattr(sys.stdin)
+
+        # File path of received commands for visualization
+        self.json_file_path = 'received_commands.json'
+    
+    def send_deal(self):
+        """ Send data after taking input from the user. """
+        termios.tcsetattr(sys.stdin, termios.TCSADRAIN, self.old_settings)
+        print("\nPlease input your commands in the format <component>,<component_id>,<command>: ", end='')
+        message = input()
+        DEFAULT_ADDRESS = 0
+        FREQ_433 = 433
+        offset_frequency = FREQ_433 - (850 if FREQ_433 > 850 else 410)
+        data = (bytes([DEFAULT_ADDRESS >> 8]) + 
+                bytes([DEFAULT_ADDRESS & 0xff]) +
+                bytes([offset_frequency]) + 
+                bytes([self.addr >> 8]) + 
+                bytes([self.addr & 0xff]) + 
+                bytes([self.offset_freq]) + 
+                message.encode())
+        self.send(data)
+        print("Message sent!")
+        self.setcbreak(sys.stdin.fileno())
+        return None
+
+    def receive_data(self):
+        data = self.receive()
+
+        # When receiving a new command
+        if data:
+            try:
+                # Split data to determine message type
+                message = data.split(",")
+
+                # For uplink commands
+                if len(message) == 3:
+                    component, component_id, command = message
+                    uplink_command = command_msg
+                    uplink_command.component = component
+                    uplink_command.component_id = int(component_id)
+                    uplink_command.command = command
+                    return uplink_command
+
+                else:
+                    return None
+            
+            except Exception as e:
+                print(f"Error handling received data: {str(e)}")
+
+    def reset_terminal_settings(self):
+        """ Reset terminal settings to their original state. """
+        termios.tcsetattr(sys.stdin, termios.TCSADRAIN, self.old_settings)
+        return None
+
+
